requirements.txt
# Core Dependencies
langchain==0.3.13
langchain-community==0.3.13
langchain-openai==0.2.14
langchain-core==0.3.28
openai==1.59.6
python-dotenv==1.0.1

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5

# Tools and Search
tavily-python==0.5.0
requests==2.32.3
beautifulsoup4==4.12.3

# CORS and Security
python-multipart==0.0.20
